refactor(engine): split poll.go into backoff.go, terminal.go, reconcile.go
Pure code motion per ADR-010: rate-limit/idle backoff math moves to
backoff.go; probe/deep-fetch terminal predicates (plus the startup
terminal scan, grouped with the tests that already exercise them in
terminal_test.go) move to terminal.go; the reconcile ticker and layer-1
status refresh move to reconcile.go; self-upgrade logic merges into the
existing upgrade.go; git-meta/advisory startup checks merge into the
existing startup.go. poll.go itself now holds only Run/poll/dispatch and
the merge-train batch routing. No logic changes.
