+++ engine/poll.go
@@ -2,7 +2,6 @@ package engine
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -19,135 +18,9 @@ import (
 	"github.com/handarbeit/fabrik/internal/itemstate"
 	"github.com/handarbeit/fabrik/stages"
 	"github.com/handarbeit/fabrik/tui"
-	"github.com/handarbeit/fabrik/warnings"
 )
 
 const idleUpgradeThreshold = 2 // consecutive idle polls before checking for upgrades
-
-// rateLimitBackoffThreshold is the fraction of GraphQL rate limit remaining
-// below which the engine activates poll backoff and logs a warning.
-const rateLimitBackoffThreshold = 0.20
-
-// rateLimitHealthyThreshold is the fraction of GraphQL rate limit remaining
-// above which the engine clears an active rate-limit backoff. Using a higher
-// threshold than rateLimitBackoffThreshold (hysteresis) prevents thrashing on
-// busy boards where quota fluctuates near the activation point.
-const rateLimitHealthyThreshold = 0.50
-
-// rateLimitMaxBackoffMultiplier caps the backoff interval as a multiple of the
-// configured poll interval (e.g. 10× = 10 * PollSeconds).
-const rateLimitMaxBackoffMultiplier = 10
-
-// rateLimitNearZeroPercent is the threshold (as a percentage of Limit) below
-// which the previous remaining count is considered "near zero" for the
-// recovery wake signal. Budget transitions from near-zero to healthy happen
-// at the hourly reset boundary — not via gradual organic replenishment — so
-// this guard prevents spurious wakes from ordinary hysteresis crossings.
-const rateLimitNearZeroPercent = 1
-
-// maxIdleBackoff is the absolute maximum poll interval during idle backoff,
-// regardless of the configured poll interval.
-const maxIdleBackoff = 5 * time.Minute
-
-// idleBackoffMultiplier returns the backoff multiplier for the given idle duration.
-// Schedule: 0–5min → 1x, 5–10min → 2x, 10–20min → 4x, 20+ min → 0 (use maxIdleBackoff).
-func idleBackoffMultiplier(idleDuration time.Duration) int {
-	switch {
-	case idleDuration < 5*time.Minute:
-		return 1
-	case idleDuration < 10*time.Minute:
-		return 2
-	case idleDuration < 20*time.Minute:
-		return 4
-	default:
-		return 0
-	}
-}
-
-// nextRateLimitLow applies two-threshold hysteresis to the rate-limit backoff state.
-// Activate when ratio < rateLimitBackoffThreshold (20%) and not already low.
-// Clear when ratio > rateLimitHealthyThreshold (50%) and currently low.
-// Between the two thresholds, state is unchanged (sticky).
-func nextRateLimitLow(current bool, ratio float64) bool {
-	if !current && ratio < rateLimitBackoffThreshold {
-		return true
-	}
-	if current && ratio > rateLimitHealthyThreshold {
-		return false
-	}
-	return current
-}
-
-// isRateLimitNearZero reports whether remaining is at or near zero relative to
-// limit (within rateLimitNearZeroPercent). Returns false when limit is 0.
-func isRateLimitNearZero(remaining, limit int) bool {
-	return limit > 0 && remaining*100 <= limit*rateLimitNearZeroPercent
-}
-
-// effectiveIdleCap returns the idle backoff cap based on webhook stream health.
-// When the webhook stream is healthy or starting up, the cap is extended to
-// webhookIdleCap (60 min) since the stream covers events that would otherwise
-// require frequent polling. Falls back to maxIdleBackoff (5 min) when unhealthy.
-func effectiveIdleCap(webhookHealthy bool) time.Duration {
-	if webhookHealthy {
-		return webhookIdleCap
-	}
-	return maxIdleBackoff
-}
-
-// computeEffectiveInterval returns the effective poll interval considering both
-// idle backoff and rate-limit backoff. The result is max(idle, rateLimit).
-// The idle component is capped at effectiveIdleCap(webhookHealthy); the rate-limit
-// component uses its own cap (rateLimitMaxBackoffMultiplier × configured).
-//
-// rateLimitRatio is the remaining-to-total GraphQL quota fraction. Pass 1.0 when
-// no rate-limit backoff is active; pass the actual fraction when backoff is active.
-// The stepwise escalation schedule (activates when ratio < 1.0):
-//
-//	>=10% remaining: 2× configured  (includes sticky hysteresis zone 20%–50%)
-//	>=5% and <10%:   4× configured
-//	>=1% and <5%:    6× configured
-//	    <1%:        10× configured  (rateLimitMaxBackoffMultiplier)
-func computeEffectiveInterval(configuredInterval time.Duration, idleDuration time.Duration, rateLimitRatio float64, webhookHealthy bool) time.Duration {
-	cap := effectiveIdleCap(webhookHealthy)
-
-	var idleInterval time.Duration
-	mult := idleBackoffMultiplier(idleDuration)
-	if mult == 0 {
-		idleInterval = cap
-	} else {
-		idleInterval = configuredInterval * time.Duration(mult)
-	}
-	if idleInterval > cap {
-		idleInterval = cap
-	}
-
-	rateLimitInterval := configuredInterval
-	if rateLimitRatio < 1.0 {
-		var rlMult int
-		switch {
-		case rateLimitRatio >= 0.10:
-			rlMult = 2
-		case rateLimitRatio >= 0.05:
-			rlMult = 4
-		case rateLimitRatio >= 0.01:
-			rlMult = 6
-		default:
-			rlMult = rateLimitMaxBackoffMultiplier
-		}
-		rateLimitInterval = configuredInterval * time.Duration(rlMult)
-		maxRL := configuredInterval * time.Duration(rateLimitMaxBackoffMultiplier)
-		if rateLimitInterval > maxRL {
-			rateLimitInterval = maxRL
-		}
-	}
-
-	effective := idleInterval
-	if rateLimitInterval > effective {
-		effective = rateLimitInterval
-	}
-	return effective
-}
 
 // isTTY reports whether stdout is connected to a terminal.
 var isTTY = func() bool {
@@ -1625,771 +1498,4 @@ func gitRevParse(dir, ref string) (string, error) {
 		return "", err
 	}
 	return strings.TrimSpace(string(out)), nil
-}
-
-// runProbeAndDeepFetch performs the per-poll probe-driven cache refresh for a
-// bootstrapped CacheImpl. It calls ProbeProjectBoard to get a minimal per-item
-// snapshot (no labels; single linked-PR node) and uses effectiveUpdatedAt to
-// detect which items need a full FetchItemDetails call. Items that haven't
-// changed since the last deep-fetch skip all GitHub traffic. This replaces the
-// previous Reconcile(shallowBoard) path and reduces GraphQL cost ~5-10x on
-// idle boards.
-//
-// Probe mutations must NOT touch the Labels field; use ProbeBoardItemUpdated,
-// not ShallowBoardItemUpdated, to preserve webhook-driven label state.
-func (e *Engine) runProbeAndDeepFetch(cacheImpl *boardcache.CacheImpl) {
-	probeItems, _, err := e.client.ProbeProjectBoard(e.cfg.Owner, e.cfg.Repo, e.cfg.ProjectNum, e.cfg.OwnerType)
-	if err != nil {
-		_, graphqlStats := e.client.RateLimitStats()
-		if graphqlStats.Limit > 0 && (graphqlStats.Remaining == 0 || float64(graphqlStats.Remaining)/float64(graphqlStats.Limit) < rateLimitBackoffThreshold) {
-			retryStr := "retrying soon"
-			if !graphqlStats.Reset.IsZero() && graphqlStats.Reset.After(time.Now()) {
-				retryStr = fmt.Sprintf("retrying after %s (local)", graphqlStats.Reset.Local().Format("15:04"))
-			}
-			e.logf(0, "warn", "rate limited — polling suspended, %s: %v\n", retryStr, err)
-			e.emitStructural(tui.RateLimitAlertEvent{Exhausted: true, Reset: graphqlStats.Reset})
-		} else {
-			e.logf(0, "cache", "probe refresh failed (using prior cache state): %v\n", err)
-		}
-		return
-	}
-
-	// Guard: 0 probe items with a populated cache indicates a transient indexer
-	// hiccup rather than a genuine board wipe. Skip this cycle; retry on the next
-	// poll. (probeProjectBoardOnce already retries 3x, but a degraded response can
-	// still slip through.)
-	if len(probeItems) == 0 && cacheImpl.IsBootstrapped() {
-		e.logf(0, "cache", "probe returned 0 items while cache has data — skipping (transient indexer hiccup)\n")
-		return
-	}
-
-	newKeys := make(map[string]bool, len(probeItems))
-	var deepFetched int
-
-	for _, pi := range probeItems {
-		repo := pi.Repo
-		if repo == "" {
-			repo = e.defaultRepo()
-		}
-		newKeys[fmt.Sprintf("%s#%d", repo, pi.Number)] = true
-
-		// Stage-membership guard: whether the item's column has a matching Fabrik stage.
-		// The guard must come after newKeys[key]=true so unconfigured items are not
-		// falsely evicted from the store by the post-loop tombstoning pass (lines below).
-		configuredStage := stages.FindStage(e.cfg.Stages, pi.Status) != nil
-
-		snap, snapErr := e.store.Get(repo, pi.Number)
-		if snapErr != nil {
-			// New item on board — skip entirely if not in a configured stage.
-			if !configuredStage {
-				continue
-			}
-			// Seed minimal state into the store.
-			minimal := gh.ProjectItem{
-				ID:             pi.ContentID,
-				ItemID:         pi.ItemID,
-				Number:         pi.Number,
-				IsPR:           pi.IsPR,
-				IsClosed:       pi.IsClosed,
-				Status:         pi.Status,
-				Repo:           repo,
-				UpdatedAt:      pi.EffectiveUpdatedAt,
-				LinkedPRNumber: pi.LinkedPRNumber,
-			}
-			e.store.Apply(itemstate.IssueOpened{Item: minimal})
-			// Probe-only terminal short-circuit: closed items in a cleanup stage whose
-			// worktree is absent have no remaining Fabrik work; skip the deep-fetch.
-			// isProbeOnlyTerminal logs the worktree-presence decision internally (FR-6).
-			if e.isProbeOnlyTerminal(minimal) {
-				e.store.Apply(itemstate.TerminalFlagSet{Repo: repo, Number: pi.Number, Terminal: true})
-				e.logf(pi.Number, "cache", "probe: new item seeded terminal — skipping deep-fetch\n")
-				continue
-			}
-			e.logf(pi.Number, "cache", "probe: new item discovered — deep-fetching\n")
-			if fetchErr := e.readClient.FetchItemDetails(&minimal); fetchErr != nil {
-				e.logf(pi.Number, "warn", "probe: deep-fetch for new item failed: %v\n", fetchErr)
-				e.store.Apply(itemstate.DeepFetchFailed{Repo: repo, Number: pi.Number, At: time.Now()})
-			} else {
-				deepFetched++
-			}
-			continue
-		}
-
-		// Detect linkage drift: probe found a different linked PR than the cache holds.
-		s := snap.State()
-		cachedPRNum := 0
-		if s.LinkedPR != nil {
-			cachedPRNum = s.LinkedPR.Number
-		}
-		if pi.LinkedPRNumber != cachedPRNum {
-			if s.LastDeepFetchAt.IsZero() {
-				// Never deep-fetched: no prior deep cache to invalidate.
-				// Treat the probe's value as authoritative — write it into LinkedPR.Number
-				// and update the prToKey reverse index without firing DeepFetchInvalidated.
-				// This suppresses spurious invalidations that occur when the bootstrapped
-				// LinkedPR.Number differs from the probe's value (e.g., cold-start with
-				// the old FetchProjectBoard path that did not populate LinkedPR.Number).
-				// Apply even when pi.LinkedPRNumber==0 to clear a stale prToKey entry
-				// if the PR was delinked between bootstrap and the first probe cycle.
-				e.store.Apply(itemstate.PRDetailsUpdated{
-					Repo:     repo,
-					Number:   pi.Number,
-					PRNumber: pi.LinkedPRNumber,
-				})
-			} else {
-				// Warm cache (has been deep-fetched): real linkage drift — invalidate.
-				e.logf(pi.Number, "cache", "probe: linkage drift (was PR #%d, now PR #%d) — invalidating deep cache\n",
-					cachedPRNum, pi.LinkedPRNumber)
-				e.store.Apply(itemstate.DeepFetchInvalidated{Repo: repo, Number: pi.Number})
-			}
-		}
-
-		// Terminal skip: if this item was previously identified as terminal and the
-		// probe still shows it in the same cleanup stage, skip deep-fetch entirely —
-		// external activity on a closed Done item has no bearing on Fabrik's work.
-		// Must run BEFORE ProbeBoardItemUpdated so we read the cached Terminal flag;
-		// applyProbeItem would clear it on status change before we could react to pi.Status.
-		// pi.Status == s.Status guards against items that moved between two cleanup stages:
-		// in that case we must apply the probe update so the store reflects the new status.
-		if s.Terminal {
-			if pst := stages.FindStage(e.cfg.Stages, pi.Status); pst != nil && pst.CleanupWorktree && pi.Status == s.Status {
-				continue // still terminal in the same cleanup stage — no deep-fetch needed
-			}
-			// Status changed (left the cleanup stage or moved to a different one) — clear
-			// the flag and fall through to normal probe processing.
-			e.store.Apply(itemstate.TerminalFlagSet{Repo: repo, Number: pi.Number, Terminal: false})
-			e.logf(pi.Number, "poll", "terminal flag cleared (status drifted to %q)\n", pi.Status)
-		}
-
-		// Apply probe state (updates IsClosed, State, IsPR, Status, UpdatedAt;
-		// intentionally skips Labels to preserve webhook-driven label state).
-		e.store.Apply(itemstate.ProbeBoardItemUpdated{Repo: repo, Number: pi.Number, Item: pi})
-
-		// Unconditionally write the probe's head SHA whenever present — the probe
-		// response is always authoritative, including for cache-fresh items that
-		// skip the deep-fetch below. This is the primary poll-mode path for keeping
-		// HeadSHA populated without relying solely on the REST FetchLinkedPR fallback.
-		if pi.LinkedPRHeadSHA != "" && pi.LinkedPRNumber != 0 {
-			e.store.Apply(itemstate.PRHeadSHAUpdated{
-				Repo:        repo,
-				Number:      pi.Number,
-				LinkedPRNum: pi.LinkedPRNumber,
-				SHA:         pi.LinkedPRHeadSHA,
-			})
-		}
-
-		// Existing item in unconfigured column: probe state updated above (keeps
-		// Status current for TUI and itemMayNeedWork), but deep-fetch is wasted work.
-		if !configuredStage {
-			continue
-		}
-
-		// Deep-fetch when cache is stale relative to effectiveUpdatedAt.
-		if cacheImpl.IsItemCacheFresh(repo, pi.Number, pi.EffectiveUpdatedAt) {
-			continue
-		}
-		e.logf(pi.Number, "cache", "probe: stale — deep-fetching\n")
-		minimal := gh.ProjectItem{
-			ID:        pi.ContentID,
-			ItemID:    pi.ItemID,
-			Number:    pi.Number,
-			IsPR:      pi.IsPR,
-			IsClosed:  pi.IsClosed,
-			Status:    pi.Status,
-			Repo:      repo,
-			UpdatedAt: pi.EffectiveUpdatedAt,
-		}
-		if fetchErr := e.readClient.FetchItemDetails(&minimal); fetchErr != nil {
-			e.logf(pi.Number, "warn", "probe: deep-fetch for stale item failed: %v\n", fetchErr)
-			e.store.Apply(itemstate.DeepFetchFailed{Repo: repo, Number: pi.Number, At: time.Now()})
-			continue
-		}
-		deepFetched++
-		// After a successful deep-fetch, check if this item is now terminal.
-		if isTerminalPredicate(minimal.Labels, minimal.Status, e.cfg.Stages) {
-			if !s.Terminal {
-				e.logf(pi.Number, "poll", "terminal flag set\n")
-			}
-			e.store.Apply(itemstate.TerminalFlagSet{Repo: repo, Number: pi.Number, Terminal: true})
-		}
-	}
-
-	// Remove items no longer on the board.
-	for _, snap := range e.store.All() {
-		key := fmt.Sprintf("%s#%d", snap.Repo(), snap.Number())
-		if !newKeys[key] {
-			e.logf(snap.Number(), "cache", "probe: item gone from board — removing from store\n")
-			e.store.Remove(snap.Repo(), snap.Number())
-		}
-	}
-
-	if deepFetched > 0 {
-		e.logf(0, "cache", "probe: deep-fetched %d item(s)\n", deepFetched)
-	}
-}
-
-// isTerminalPredicate reports whether an item with the given labels and board
-// status qualifies as terminal: it must be in a cleanup (Done) stage, carry the
-// stage:Name:complete label, and have no transient lifecycle or lock labels.
-// Terminal items have no remaining Fabrik work and may safely skip deep-fetch.
-func isTerminalPredicate(labels []string, status string, stagesCfg []*stages.Stage) bool {
-	st := stages.FindStage(stagesCfg, status)
-	if st == nil || !st.CleanupWorktree {
-		return false
-	}
-	completeLabel := "stage:" + st.Name + ":complete"
-	hasComplete := false
-	for _, l := range labels {
-		if l == completeLabel {
-			hasComplete = true
-			break
-		}
-	}
-	if !hasComplete {
-		return false
-	}
-	for _, l := range labels {
-		for _, tl := range transientLifecycleLabels {
-			if l == tl {
-				return false
-			}
-		}
-		if strings.HasPrefix(l, "fabrik:locked:") {
-			return false
-		}
-	}
-	return true
-}
-
-// isProbeOnlyTerminal reports whether an item is terminal based on probe data
-// (IsClosed + CleanupWorktree stage) and confirms the on-disk worktree is absent.
-// Use this predicate in the new-item branch of runProbeAndDeepFetch and in
-// seedTerminalFromProbeItems, where labels have not yet been fetched.
-//
-// The worktree check prevents stranding cleanup work: if the worktree still
-// exists on disk (cleanup_worktree stage not yet run), the item must proceed
-// through processItem so the Done stage cleanup can execute.
-func (e *Engine) isProbeOnlyTerminal(item gh.ProjectItem) bool {
-	if !item.IsClosed {
-		return false
-	}
-	st := stages.FindStage(e.cfg.Stages, item.Status)
-	if st == nil || !st.CleanupWorktree {
-		return false
-	}
-	if e.worktreeExistsForItem(item) {
-		e.logf(item.Number, "cache", "probe: worktree present — not treating as terminal yet\n")
-		return false
-	}
-	e.logf(item.Number, "cache", "probe: no worktree on disk — treating as terminal\n")
-	return true
-}
-
-// seedTerminalFromProbeItems applies TerminalFlagSet for probe items that
-// qualify as terminal: closed, in a cleanup stage, and worktree absent on disk.
-// Must be called after BootstrapFromProbe so items are in the store.
-func (e *Engine) seedTerminalFromProbeItems(items []gh.BoardProbeItem) {
-	var seeded int
-	for _, pi := range items {
-		stub := gh.ProjectItem{
-			Number:   pi.Number,
-			IsClosed: pi.IsClosed,
-			Status:   pi.Status,
-			Repo:     pi.Repo,
-		}
-		if e.isProbeOnlyTerminal(stub) {
-			e.store.Apply(itemstate.TerminalFlagSet{
-				Repo:     pi.Repo,
-				Number:   pi.Number,
-				Terminal: true,
-			})
-			seeded++
-		}
-	}
-	if seeded > 0 {
-		e.logf(0, "cache", "probe bootstrap: seeded %d terminal items (worktree absent, closed cleanup stage)\n", seeded)
-	}
-}
-
-// runStartupTransientLabelScan is a one-shot recovery pass that runs after the
-// first successful poll. It scans the Store for closed issues that still carry
-// transient lifecycle labels — a condition that can occur when an issue closes
-// mid-stage during a prior crash.
-//
-// Label availability note: when the cache was bootstrapped via BootstrapFromProbe
-// (the default cold-start path), labels are absent from the Store and this scan
-// is a no-op. The accepted gap: stale transient labels on closed terminal items
-// will not be cleaned up at startup (very low probability — requires a crash
-// between issue close and label removal in the Done stage). Active items are
-// deep-fetched on the first probe cycle, which populates their labels normally,
-// so those items are covered by the steady-state cleanup path.
-// When bootstrapped via the full FetchProjectBoard path (webhook-paused / reconcile),
-// the Store contains full label data and this scan is fully effective.
-func (e *Engine) runStartupTransientLabelScan() {
-	snaps := e.store.All()
-	if len(snaps) == 0 {
-		return
-	}
-
-	// Build a synthetic board containing only closed items that carry at least
-	// one transient lifecycle label or a lock label. The cleanup helpers operate
-	// on *gh.ProjectBoard items so we pass only the relevant subset.
-	transientSet := make(map[string]bool, len(transientLifecycleLabels))
-	for _, l := range transientLifecycleLabels {
-		transientSet[l] = true
-	}
-	lockLabel := fmt.Sprintf("fabrik:locked:%s", e.cfg.User)
-
-	var items []gh.ProjectItem
-	for _, snap := range snaps {
-		if !snap.IsClosed() {
-			continue
-		}
-		labels := snap.Labels()
-		hasStale := false
-		for _, l := range labels {
-			if transientSet[l] || l == lockLabel {
-				hasStale = true
-				break
-			}
-		}
-		if !hasStale {
-			continue
-		}
-		items = append(items, gh.ProjectItem{
-			Number:   snap.Number(),
-			Repo:     snap.Repo(),
-			IsClosed: true,
-			Labels:   labels,
-		})
-	}
-	if len(items) == 0 {
-		return
-	}
-	e.logf(0, "startup", "transient-label scan: %d closed item(s) with stale labels\n", len(items))
-	board := &gh.ProjectBoard{Items: items}
-	e.cleanupClosedIssueLocks(board)
-	e.cleanupClosedIssueTransientLabels(board)
-}
-
-// runStartupTerminalScan marks terminal items in the Store after the first
-// successful poll using the full label-aware predicate (isTerminalPredicate).
-// Must run after runStartupTransientLabelScan so stale transient labels have
-// been removed before the predicate is evaluated.
-//
-// Label availability note: when the cache was bootstrapped via BootstrapFromProbe
-// (the default cold-start path), labels are absent from the Store. In that case
-// this scan is a no-op — isTerminalPredicate returns false for every item.
-// Cold-start terminal seeding is instead handled by seedTerminalFromProbeItems
-// (called after BootstrapFromProbe) using IsClosed+CleanupWorktree+worktree-absent.
-// When bootstrapped via the full FetchProjectBoard path, labels are present
-// and this scan applies the full predicate correctly.
-func (e *Engine) runStartupTerminalScan() {
-	snaps := e.store.All()
-	var marked int
-	for _, snap := range snaps {
-		if snap.IsTerminal() {
-			continue // already set — no-op path
-		}
-		if isTerminalPredicate(snap.Labels(), snap.Status(), e.cfg.Stages) {
-			e.store.Apply(itemstate.TerminalFlagSet{
-				Repo:     snap.Repo(),
-				Number:   snap.Number(),
-				Terminal: true,
-			})
-			marked++
-		}
-	}
-	if marked > 0 {
-		e.logf(0, "startup", "terminal scan: marked %d item(s) as terminal\n", marked)
-	}
-}
-
-// checkAndUpgrade selects the upgrade path based on the running version:
-//   - dev builds (version starts with "dev"): git pull → go build → re-exec
-//   - release builds (all other versions): GitHub Releases API → download → atomic replace → re-exec
-func (e *Engine) checkAndUpgrade() {
-	if !strings.HasPrefix(e.cfg.Version, "dev") {
-		e.checkReleaseUpgrade()
-		return
-	}
-
-	dir := e.fabrikDir
-
-	// Only auto-upgrade if we're running from a Fabrik source checkout.
-	if !isFabrikSourceCheckout(dir) {
-		return
-	}
-
-	baseBranch := "main"
-
-	// Check local HEAD first — detects local commits that haven't been pushed.
-	localRef, err := gitRevParse(dir, "HEAD")
-	if err != nil {
-		e.logf(0, "upgrade", "could not resolve HEAD: %v\n", err)
-		return
-	}
-	binarySHA := extractBinarySHA(e.cfg.Version)
-	needsRebuild := binarySHA != "" && !strings.HasPrefix(localRef, binarySHA)
-	if needsRebuild {
-		e.logf(0, "upgrade", "binary built from %s but HEAD is %s — rebuilding\n", binarySHA, localRef[:7])
-	}
-
-	// Also check remote for new upstream commits.
-	if !needsRebuild {
-		pollStatus("[upgrade] checking origin/%s ...", baseBranch)
-
-		fetchCmd := exec.Command("git", "fetch", "origin", baseBranch)
-		fetchCmd.Dir = dir
-		if out, err := fetchCmd.CombinedOutput(); err != nil {
-			e.logf(0, "upgrade", "git fetch failed: %v\n%s\n", err, out)
-			return
-		}
-
-		remoteRef, err := gitRevParse(dir, "origin/"+baseBranch)
-		if err != nil {
-			e.logf(0, "upgrade", "could not resolve origin/%s: %v\n", baseBranch, err)
-			return
-		}
-		if localRef == remoteRef {
-			pollStatusClear()
-			return // up to date
-		}
-		// Only pull if remote is ahead of local. If local is ahead (unpushed
-		// commits), we already checked the binary SHA against local HEAD above.
-		mergeBaseCmd := exec.Command("git", "merge-base", "--is-ancestor", localRef, remoteRef)
-		mergeBaseCmd.Dir = dir
-		if err := mergeBaseCmd.Run(); err != nil {
-			// localRef is not an ancestor of remoteRef — local is ahead or diverged.
-			// Either way, nothing to pull. The binary SHA check above already
-			// handled whether a rebuild is needed.
-			pollStatusClear()
-			return
-		}
-		needsRebuild = true
-		e.logf(0, "upgrade", "new commits on origin/%s — pulling\n", baseBranch)
-
-		pullCmd := exec.Command("git", "pull", "--ff-only", "origin", baseBranch)
-		pullCmd.Dir = dir
-		if out, err := pullCmd.CombinedOutput(); err != nil {
-			e.logf(0, "upgrade", "git pull --ff-only failed (local changes?): %v\n%s\n", err, out)
-			return
-		}
-	}
-
-	exe, err := os.Executable()
-	if err != nil {
-		e.logf(0, "upgrade", "could not determine executable path: %v\n", err)
-		return
-	}
-	exe, err = filepath.EvalSymlinks(exe)
-	if err != nil {
-		e.logf(0, "upgrade", "could not resolve symlinks for executable: %v\n", err)
-		return
-	}
-
-	e.logf(0, "upgrade", "rebuilding binary: %s\n", exe)
-
-	buildCmd := exec.Command("go", "build", "-o", exe, ".")
-	buildCmd.Dir = dir
-	if out, err := buildCmd.CombinedOutput(); err != nil {
-		e.logf(0, "upgrade", "build failed: %v\n%s\n", err, out)
-		return
-	}
-
-	// Refresh plugin skills from the new binary.
-	e.logf(0, "upgrade", "refreshing plugin skills\n")
-	upgradeCmd := exec.Command(exe, "upgrade")
-	upgradeCmd.Dir = dir
-	if out, err := upgradeCmd.CombinedOutput(); err != nil {
-		e.logf(0, "upgrade", "fabrik upgrade failed: %v\n%s\n", err, out)
-		// Non-fatal — continue with re-exec, old skills still work
-	}
-
-	e.logf(0, "upgrade", "re-executing new binary\n")
-
-	if err := syscall.Exec(exe, os.Args, os.Environ()); err != nil {
-		e.logf(0, "upgrade", "exec failed: %v\n", err)
-	}
-}
-
-// extractBinarySHA extracts the short SHA from a dev version string like
-// "dev(abc1234)". Returns "" if the version is not a dev build or has no SHA.
-func extractBinarySHA(version string) string {
-	if !strings.HasPrefix(version, "dev(") || !strings.HasSuffix(version, ")") {
-		return ""
-	}
-	return version[4 : len(version)-1]
-}
-
-// isFabrikSourceCheckout reports whether dir is a git checkout of the fabrik
-// source repo (handarbeit/fabrik). Returns false on any error (no git, no
-// remote, wrong remote, etc.).
-func isFabrikSourceCheckout(dir string) bool {
-	cmd := exec.Command("git", "remote", "get-url", "origin")
-	cmd.Dir = dir
-	out, err := cmd.Output()
-	if err != nil {
-		return false
-	}
-	url := strings.TrimSuffix(strings.TrimSpace(string(out)), ".git")
-	return strings.Contains(url, "handarbeit/fabrik")
-}
-
-// checkReleaseUpgrade is the release-based upgrade path. It checks the GitHub
-// Releases API for a version newer than the running binary, downloads the
-// matching platform asset, atomically replaces the running binary, and re-execs.
-//
-// All failures are non-fatal: a warning is logged and the poll loop continues.
-func (e *Engine) checkReleaseUpgrade() {
-	logf := func(format string, args ...any) {
-		e.logf(0, "upgrade", format, args...)
-	}
-	// Error discarded intentionally: failures are logged by PerformReleaseUpgrade
-	// itself via logf, and this caller's contract is non-fatal — the poll loop
-	// continues regardless (unlike the foreground `fabrik upgrade` command).
-	_ = PerformReleaseUpgrade(e.client, e.cfg.Version, e.cfg.Token, []string{"FABRIK_AUTO_UPGRADED=1"}, logf)
-}
-
-// captureGitMeta captures the current branch name, short commit SHA,
-// origin/{baseBranch} SHA, and a human-readable UTC timestamp from the given
-// worktree directory. Returns "unknown" values gracefully if git commands fail.
-// mainSHA is empty (not "unknown") when it cannot be resolved — callers treat
-// empty as "no data" rather than an error sentinel.
-func captureGitMeta(workDir, baseBranch string) (branch, commit, mainSHA, timestamp string) {
-	timestamp = time.Now().UTC().Format("2006-01-02 15:04 UTC")
-
-	if workDir == "" {
-		return "unknown", "unknown", "", timestamp
-	}
-
-	sha, err := gitRevParse(workDir, "HEAD")
-	if err != nil || sha == "" {
-		commit = "unknown"
-	} else if len(sha) >= 8 {
-		commit = sha[:8]
-	} else {
-		commit = sha
-	}
-
-	branchCmd := exec.Command("git", "rev-parse", "--abbrev-ref", "HEAD")
-	branchCmd.Dir = workDir
-	out, err := branchCmd.Output()
-	if err != nil {
-		branch = "unknown"
-	} else {
-		branch = strings.TrimSpace(string(out))
-	}
-
-	// Capture origin/{baseBranch} SHA for staleness tracking.
-	// Store full SHA — it is used as a git revision in writeCodebaseChanges;
-	// abbreviated SHAs can become ambiguous in larger repos.
-	if baseBranch != "" {
-		if mSHA, err := gitRevParse(workDir, "origin/"+baseBranch); err == nil {
-			mainSHA = mSHA
-		}
-	}
-
-	return branch, commit, mainSHA, timestamp
-}
-
-// checkHTTPSCredentials probes whether a git credential helper is configured
-// when using HTTPS clone mode. Prints an advisory warning if none is found.
-// Skip entirely when SSH mode is active or when HTTPS→SSH URL rewriting is in
-// effect for github.com — no credential helper is needed in either case.
-// This check is non-interactive and never prompts; it only reads git config.
-func (e *Engine) checkHTTPSCredentials(hasSSHRewrite bool) {
-	if e.cfg.GitSSH || hasSSHRewrite {
-		return
-	}
-	cmd := exec.Command("git", "config", "credential.helper")
-	out, err := cmd.Output()
-	if err != nil || len(strings.TrimSpace(string(out))) == 0 {
-		fmt.Printf("[startup] warn: no git credential helper configured; HTTPS cloning may prompt for credentials.\n")
-		fmt.Printf("[startup] warn: configure one (e.g. git-credential-osxkeychain) or use --ssh / git_ssh: true in .fabrik/config.yaml.\n")
-		fmt.Printf("[startup] note: existing bare clones retain their original remote URL and are not affected by this setting.\n")
-	}
-}
-
-// checkAllowAutoMerge queries the GitHub API for the allow_auto_merge setting on
-// owner/repo and prints a WARNING if it is disabled. Non-fatal: API errors are
-// logged at warn level and processing continues. The check fires at most once per
-// repo per process run (guarded by checkedAutoMergeRepos).
-func (e *Engine) checkAllowAutoMerge(owner, repo string) {
-	key := owner + "/" + repo
-	e.mu.Lock()
-	already := e.checkedAutoMergeRepos[key]
-	e.checkedAutoMergeRepos[key] = true
-	e.mu.Unlock()
-	if already {
-		return
-	}
-	enabled, err := e.client.FetchAllowAutoMerge(owner, repo)
-	if err != nil {
-		e.logf(0, "warn", "could not check allow_auto_merge for %s: %v\n", key, err)
-		return
-	}
-	if !enabled {
-		e.logf(0, "startup", "WARNING: %s has allow_auto_merge disabled.\n", key)
-		e.logf(0, "startup", "WARNING: yolo issues on this repo will reach Validate complete but their PRs will not merge.\n")
-		e.logf(0, "startup", "WARNING: Fix: gh api -X PATCH repos/%s -f allow_auto_merge=true\n", key)
-		_ = warnings.Record(warnings.Entry{
-			Key:       "allow_auto_merge:" + key,
-			Type:      "allow_auto_merge",
-			Title:     "allow_auto_merge disabled on " + key,
-			Detail:    fmt.Sprintf("yolo issues on this repo will reach Validate complete but their PRs will not merge.\n\nFix: gh api -X PATCH repos/%s -f allow_auto_merge=true", key),
-			FixAction: "shell_command",
-			FixParams: map[string]string{"cmd": fmt.Sprintf("gh api -X PATCH repos/%s -f allow_auto_merge=true", key)},
-		})
-	} else {
-		_ = warnings.Clear("allow_auto_merge:" + key)
-	}
-}
-
-// checkURLRewrite detects whether git has URL rewriting configured that
-// transparently redirects github.com HTTPS URLs to SSH (e.g. via
-// url.git@github.com:.insteadOf = https://github.com/ in ~/.gitconfig).
-// Returns true when such HTTPS→SSH rewriting is active. Prints an
-// informational notice when active — git applies the rewriting transparently,
-// so Fabrik's HTTPS clone URLs will automatically use SSH.
-func (e *Engine) checkURLRewrite() bool {
-	cmd := exec.Command("git", "config", "--get-regexp", `url\..*\.insteadOf`)
-	out, _ := cmd.Output() // exit code 1 = no matches, not an error
-	// Parse each line: "url.<base>.insteadof <value>"
-	// Look specifically for entries that rewrite https://github.com URLs to an
-	// SSH base (key contains git@github.com), avoiding false positives from
-	// same-protocol or reverse (SSH→HTTPS) rewrites.
-	for _, line := range strings.Split(strings.TrimSpace(string(out)), "\n") {
-		if line == "" {
-			continue
-		}
-		parts := strings.SplitN(line, " ", 2)
-		if len(parts) != 2 {
-			continue
-		}
-		key := strings.ToLower(parts[0])     // url.<base>.insteadof
-		value := strings.TrimSpace(parts[1]) // the insteadOf value (URL prefix to match)
-		if strings.Contains(value, "https://github.com") && strings.Contains(key, "git@github.com") {
-			fmt.Printf("[startup] note: git URL rewriting for github.com is active (HTTPS → SSH); Fabrik's HTTPS clone URLs will be transparently redirected to SSH via your git config.\n")
-			return true
-		}
-	}
-	return false
-}
-
-// reconcileLoop is the poll-only correctness backstop for cache/GitHub divergence.
-// It periodically runs LightReconcile (a fresh shallow board fetch + drift compare)
-// and, on drift, reconciles the cache — re-syncing shallow fields including the
-// fabrik-managed label set. It MUST run whether or not the webhook manager started
-// (#955): a webhook-less (or webhook-failed) deployment must still self-heal, so
-// this loop is launched unconditionally rather than nested in the webhook-start
-// path. wm may be nil (webhooks off/failed); webhook health-state transitions are
-// skipped in that case, but drift detection and repair still run.
-func (e *Engine) reconcileLoop(ctx context.Context, cacheImpl *boardcache.CacheImpl, wm *webhookManager) {
-	reconcileInterval := e.cfg.ReconcileInterval
-	if reconcileInterval <= 0 {
-		reconcileInterval = lightReconcileInterval
-	}
-	ticker := time.NewTicker(reconcileInterval)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			driftCount, driftedKeys, freshBoard, err := cacheImpl.LightReconcile(
-				e.cfg.Owner, e.cfg.Repo, e.cfg.ProjectNum, e.cfg.OwnerType,
-			)
-			if err != nil {
-				e.logf(0, "reconcile", "light reconcile failed (no health state change): %v\n", err)
-				continue
-			}
-			if driftCount == 0 {
-				if wm != nil {
-					wm.transitionHealthState(WebhookStreamHealthy, "")
-				}
-				continue
-			}
-			keyStr := fmt.Sprintf("%v", driftedKeys)
-			if len(driftedKeys) > 5 {
-				keyStr = fmt.Sprintf("%v … %d more", driftedKeys[:5], len(driftedKeys)-5)
-			}
-			e.logf(0, "reconcile", "light reconcile: %d item(s) drifted (%s) — reconciling cache\n", driftCount, keyStr)
-			if wm != nil {
-				wm.transitionHealthState(WebhookStreamUnhealthy, fmt.Sprintf("%d item(s) drifted", driftCount))
-			}
-			cacheImpl.Pause()
-			cacheImpl.Reconcile(freshBoard)
-			cacheImpl.Resume()
-			if wm != nil {
-				wm.transitionHealthState(WebhookStreamHealthy, "drift reconciled")
-			}
-		}
-	}
-}
-
-// applyLayer1StatusRefresh handles the Layer 1 opportunistic per-event Status
-// refresh. Called from the deltaFn closure after ApplyDelta. For issue and
-// issue_comment events, it fetches the current Status from GitHub and updates
-// the cache immediately. Two paths:
-//   - Fast path: cache has the item's itemID → calls FetchProjectItemStatus.
-//   - Fallback path: cache lacks itemID (brand-new issue, issues.opened before
-//     projects_v2_item.created) → calls LookupIssueProjectItem to populate
-//     both itemID and Status in one query. Skipped when cache.ProjectID() == ""
-//     (Bootstrap not yet complete).
-//
-// All errors are best-effort: logged as warnings and never returned.
-func (e *Engine) applyLayer1StatusRefresh(eventType string, payload []byte, cache *boardcache.CacheImpl) {
-	if cache.IsPaused() {
-		return
-	}
-	if eventType != "issues" && eventType != "issue_comment" {
-		return
-	}
-	var ev struct {
-		Issue struct {
-			Number int `json:"number"`
-		} `json:"issue"`
-		Repository struct {
-			FullName string `json:"full_name"`
-		} `json:"repository"`
-	}
-	if err := json.Unmarshal(payload, &ev); err != nil {
-		e.logf(0, "warn", "layer1 status refresh: failed to parse %s payload: %v\n", eventType, err)
-		return
-	}
-	if ev.Repository.FullName == "" || ev.Issue.Number == 0 {
-		return
-	}
-	key := boardcache.ItemKey(ev.Repository.FullName, ev.Issue.Number)
-	itemID, ok := cache.GetItemID(key)
-	if !ok {
-		// Fallback: issue is in the cache but has no itemID yet (e.g., arrived via
-		// issues.opened before a projects_v2_item.created event). Perform a single
-		// GraphQL lookup to populate both itemID and Status in one call.
-		// If Bootstrap hasn't completed yet, projectID is empty — skip to avoid a
-		// useless API call with an empty project ID during the startup window.
-		projectID := cache.ProjectID()
-		if projectID == "" {
-			return
-		}
-		fetchedItemID, fetchedStatus, err := e.client.LookupIssueProjectItem(projectID, ev.Repository.FullName, ev.Issue.Number)
-		if err != nil {
-			e.logf(ev.Issue.Number, "warn", "layer1 fallback lookup failed for %s#%d: %v\n", ev.Repository.FullName, ev.Issue.Number, err)
-			return
-		}
-		if fetchedItemID == "" {
-			// Issue is not on the project fabrik manages — silently skip.
-			return
-		}
-		cache.RegisterItemID(key, fetchedItemID)
-		cache.UpdateItemStatus(key, fetchedStatus)
-		return
-	}
-	status, err := e.client.FetchProjectItemStatus(itemID)
-	if err != nil {
-		e.logf(ev.Issue.Number, "warn", "layer1 status fetch failed for %s: %v\n", itemID, err)
-		return
-	}
-	cache.UpdateItemStatus(key, status)
 }
