+++ engine/upgrade.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"net/http"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"runtime"
 	"strconv"
@@ -252,4 +253,150 @@ func PerformReleaseUpgrade(client GitHubClient, version, token string, extraEnv 
 		return fmt.Errorf("re-executing upgraded binary: %w", err)
 	}
 	return nil
+}
+
+// checkAndUpgrade selects the upgrade path based on the running version:
+//   - dev builds (version starts with "dev"): git pull → go build → re-exec
+//   - release builds (all other versions): GitHub Releases API → download → atomic replace → re-exec
+func (e *Engine) checkAndUpgrade() {
+	if !strings.HasPrefix(e.cfg.Version, "dev") {
+		e.checkReleaseUpgrade()
+		return
+	}
+
+	dir := e.fabrikDir
+
+	// Only auto-upgrade if we're running from a Fabrik source checkout.
+	if !isFabrikSourceCheckout(dir) {
+		return
+	}
+
+	baseBranch := "main"
+
+	// Check local HEAD first — detects local commits that haven't been pushed.
+	localRef, err := gitRevParse(dir, "HEAD")
+	if err != nil {
+		e.logf(0, "upgrade", "could not resolve HEAD: %v\n", err)
+		return
+	}
+	binarySHA := extractBinarySHA(e.cfg.Version)
+	needsRebuild := binarySHA != "" && !strings.HasPrefix(localRef, binarySHA)
+	if needsRebuild {
+		e.logf(0, "upgrade", "binary built from %s but HEAD is %s — rebuilding\n", binarySHA, localRef[:7])
+	}
+
+	// Also check remote for new upstream commits.
+	if !needsRebuild {
+		pollStatus("[upgrade] checking origin/%s ...", baseBranch)
+
+		fetchCmd := exec.Command("git", "fetch", "origin", baseBranch)
+		fetchCmd.Dir = dir
+		if out, err := fetchCmd.CombinedOutput(); err != nil {
+			e.logf(0, "upgrade", "git fetch failed: %v\n%s\n", err, out)
+			return
+		}
+
+		remoteRef, err := gitRevParse(dir, "origin/"+baseBranch)
+		if err != nil {
+			e.logf(0, "upgrade", "could not resolve origin/%s: %v\n", baseBranch, err)
+			return
+		}
+		if localRef == remoteRef {
+			pollStatusClear()
+			return // up to date
+		}
+		// Only pull if remote is ahead of local. If local is ahead (unpushed
+		// commits), we already checked the binary SHA against local HEAD above.
+		mergeBaseCmd := exec.Command("git", "merge-base", "--is-ancestor", localRef, remoteRef)
+		mergeBaseCmd.Dir = dir
+		if err := mergeBaseCmd.Run(); err != nil {
+			// localRef is not an ancestor of remoteRef — local is ahead or diverged.
+			// Either way, nothing to pull. The binary SHA check above already
+			// handled whether a rebuild is needed.
+			pollStatusClear()
+			return
+		}
+		needsRebuild = true
+		e.logf(0, "upgrade", "new commits on origin/%s — pulling\n", baseBranch)
+
+		pullCmd := exec.Command("git", "pull", "--ff-only", "origin", baseBranch)
+		pullCmd.Dir = dir
+		if out, err := pullCmd.CombinedOutput(); err != nil {
+			e.logf(0, "upgrade", "git pull --ff-only failed (local changes?): %v\n%s\n", err, out)
+			return
+		}
+	}
+
+	exe, err := os.Executable()
+	if err != nil {
+		e.logf(0, "upgrade", "could not determine executable path: %v\n", err)
+		return
+	}
+	exe, err = filepath.EvalSymlinks(exe)
+	if err != nil {
+		e.logf(0, "upgrade", "could not resolve symlinks for executable: %v\n", err)
+		return
+	}
+
+	e.logf(0, "upgrade", "rebuilding binary: %s\n", exe)
+
+	buildCmd := exec.Command("go", "build", "-o", exe, ".")
+	buildCmd.Dir = dir
+	if out, err := buildCmd.CombinedOutput(); err != nil {
+		e.logf(0, "upgrade", "build failed: %v\n%s\n", err, out)
+		return
+	}
+
+	// Refresh plugin skills from the new binary.
+	e.logf(0, "upgrade", "refreshing plugin skills\n")
+	upgradeCmd := exec.Command(exe, "upgrade")
+	upgradeCmd.Dir = dir
+	if out, err := upgradeCmd.CombinedOutput(); err != nil {
+		e.logf(0, "upgrade", "fabrik upgrade failed: %v\n%s\n", err, out)
+		// Non-fatal — continue with re-exec, old skills still work
+	}
+
+	e.logf(0, "upgrade", "re-executing new binary\n")
+
+	if err := syscall.Exec(exe, os.Args, os.Environ()); err != nil {
+		e.logf(0, "upgrade", "exec failed: %v\n", err)
+	}
+}
+
+// extractBinarySHA extracts the short SHA from a dev version string like
+// "dev(abc1234)". Returns "" if the version is not a dev build or has no SHA.
+func extractBinarySHA(version string) string {
+	if !strings.HasPrefix(version, "dev(") || !strings.HasSuffix(version, ")") {
+		return ""
+	}
+	return version[4 : len(version)-1]
+}
+
+// isFabrikSourceCheckout reports whether dir is a git checkout of the fabrik
+// source repo (handarbeit/fabrik). Returns false on any error (no git, no
+// remote, wrong remote, etc.).
+func isFabrikSourceCheckout(dir string) bool {
+	cmd := exec.Command("git", "remote", "get-url", "origin")
+	cmd.Dir = dir
+	out, err := cmd.Output()
+	if err != nil {
+		return false
+	}
+	url := strings.TrimSuffix(strings.TrimSpace(string(out)), ".git")
+	return strings.Contains(url, "handarbeit/fabrik")
+}
+
+// checkReleaseUpgrade is the release-based upgrade path. It checks the GitHub
+// Releases API for a version newer than the running binary, downloads the
+// matching platform asset, atomically replaces the running binary, and re-execs.
+//
+// All failures are non-fatal: a warning is logged and the poll loop continues.
+func (e *Engine) checkReleaseUpgrade() {
+	logf := func(format string, args ...any) {
+		e.logf(0, "upgrade", format, args...)
+	}
+	// Error discarded intentionally: failures are logged by PerformReleaseUpgrade
+	// itself via logf, and this caller's contract is non-fatal — the poll loop
+	// continues regardless (unlike the foreground `fabrik upgrade` command).
+	_ = PerformReleaseUpgrade(e.client, e.cfg.Version, e.cfg.Token, []string{"FABRIK_AUTO_UPGRADED=1"}, logf)
 }
